MAINT: Move spx logic into utils
